use clippy for lint

+++ src/debugger/metrics/metrics.rs
@@ -28,7 +28,7 @@ impl<T> MetricsExport<T> {
 
     #[allow(dead_code)]
     pub fn secs_ratio(&self) -> f64 {
-        1 as f64 / self.duration.as_secs_f64()
+        1f64 / self.duration.as_secs_f64()
     }
 }
